feat: add basic ui elements for barcode overlay and scanner overlay, update scanWindow via MobileScanner widget
