+++ lib/mobile_scanner.dart
@@ -27,3 +27,7 @@ export 'src/objects/phone.dart';
 export 'src/objects/sms.dart';
 export 'src/objects/url_bookmark.dart';
 export 'src/objects/wifi.dart';
+export 'src/overlay/barcode_overlay.dart';
+export 'src/overlay/barcode_painter.dart';
+export 'src/overlay/scanner_overlay.dart';
+export 'src/overlay/scanner_painter.dart';
