+++ lib/src/overlay/barcode_painter.dart
@@ -1,0 +1,67 @@
+import 'package:flutter/material.dart';
+
+class BarcodePainter extends CustomPainter {
+  BarcodePainter({
+    required this.barcodeCorners,
+    required this.barcodeSize,
+    required this.boxFit,
+    required this.cameraPreviewSize,
+  });
+
+  final List<Offset> barcodeCorners;
+  final Size barcodeSize;
+  final BoxFit boxFit;
+  final Size cameraPreviewSize;
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    if (barcodeCorners.isEmpty ||
+        barcodeSize.isEmpty ||
+        cameraPreviewSize.isEmpty) {
+      return;
+    }
+
+    final adjustedSize = applyBoxFit(boxFit, cameraPreviewSize, size);
+
+    double verticalPadding = size.height - adjustedSize.destination.height;
+    double horizontalPadding = size.width - adjustedSize.destination.width;
+    if (verticalPadding > 0) {
+      verticalPadding = verticalPadding / 2;
+    } else {
+      verticalPadding = 0;
+    }
+
+    if (horizontalPadding > 0) {
+      horizontalPadding = horizontalPadding / 2;
+    } else {
+      horizontalPadding = 0;
+    }
+
+    final double ratioWidth;
+    final double ratioHeight;
+
+    ratioWidth = cameraPreviewSize.width / adjustedSize.destination.width;
+    ratioHeight = cameraPreviewSize.height / adjustedSize.destination.height;
+
+    final List<Offset> adjustedOffset = [
+      for (final offset in barcodeCorners)
+        Offset(
+          offset.dx / ratioWidth + horizontalPadding,
+          offset.dy / ratioHeight + verticalPadding,
+        ),
+    ];
+
+    final cutoutPath = Path()..addPolygon(adjustedOffset, true);
+
+    final backgroundPaint = Paint()
+      ..color = Colors.red.withOpacity(0.3)
+      ..style = PaintingStyle.fill;
+
+    canvas.drawPath(cutoutPath, backgroundPaint);
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) {
+    return true;
+  }
+}
